requirements.txt
langchain
python-dotenv
fschat
chromadb
sentence_transformers
InstructorEmbedding
gradio
bs4
protobuf==3.20
openai
tiktoken